requirements.txt
geopandas
rasterio
pyproj
numpy
pandas
matplotlib
scikit-learn
torch
torchvision
torchmetrics
leafmap
Pillow